fat trimmed. now refactor
